# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      detail:
        description: optional internal error
        example: validation failed on field 'email'
        type: string
      error:
        description: user-friendly message
        example: Invalid request
        type: string
    type: object
...
  order.CreateOrderRequestDoc:
    properties:
      admin_id:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      delivery_address:
        type: string
      delivery_point:
        $ref: '#/definitions/order.Point'
      image_url:
        type: string
      merchant_id:
        type: string
      pickup_address:
        type: string
      pickup_point:
        $ref: '#/definitions/order.Point'
      product_id:
        type: string
      product_name:
        description: snapshot at purchase time
        type: string
      quantity:
        type: integer
      total:
        type: integer
      unit_price:
        type: integer
      variant_id:
        type: string
      variant_name:
        type: string
    required:
    - admin_id
    - currency
    - customer_id
    - delivery_address
    - delivery_point
    - image_url
    - merchant_id
    - pickup_address
    - pickup_point
    - product_id
    - product_name
    - quantity
    - total
    - unit_price
    - variant_id
    - variant_name
    type: object
  order.OrderDoc:
    properties:
      admin_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      delivery_address:
        type: string
      delivery_point:
        $ref: '#/definitions/order.Point'
      id:
        type: string
      image_url:
        type: string
      merchant_id:
        type: string
      pickup_address:
        type: string
      pickup_point:
        $ref: '#/definitions/order.Point'
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      status:
        $ref: '#/definitions/order.OrderStatus'
      total:
        type: integer
      unit_price:
        type: integer
      updated_at:
        type: string
      variant_id:
        type: string
      variant_name:
        type: string
    type: object
  order.OrderStatus:
    enum:
    - pending
    - assigned
    - in_transit
    - delivered
    - cancelled
    type: string
    x-enum-varnames:
    - Pending
    - Assigned
    - InTransit
    - Delivered
    - Cancelled
  order.Point:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
  order.UpdateOrderRequest:
    properties:
      column:
        description: e.g. "status", "quantity"
        type: string
      value:
        description: Accepts string, int, etc.
    required:
    - column
    - value
    type: object
  payment.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      currency:
        type: string
      method:
        $ref: '#/definitions/payment.PaymentMethod'
      order_id:
        type: string
      status:
        $ref: '#/definitions/payment.PaymentStatus'
    type: object
  payment.Payment:
    properties:
      amount:
        type: integer
      currency:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/payment.PaymentMethod'
      order_id:
        type: string
      paid_at:
        type: string
      status:
        $ref: '#/definitions/payment.PaymentStatus'
    type: object
  payment.PaymentMethod:
    enum:
    - stripe
    - paypal
    - mobile_money
    - cash_on_delivery
    type: string
    x-enum-varnames:
    - MethodStripe
    - MethodPayPal
    - MethodMobileMoney
    - MethodCashOnDelivery
  payment.PaymentStatus:
    enum:
    - pending
    - completed
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusCompleted
    - StatusFailed
...
host: localhost:8000
info:
  contact: {}
  description: This is the API for fastabiz operations.
  title: FastaBiz API
  version: "1.0"
paths:
...
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes an order by their ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a order
      tags:
      - orders
  /orders/{id}/update:
    put:
      consumes:
      - application/json
      description: Update any order struct field of an existing order
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      - description: Field and value to update
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/order.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid order ID or request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Order
      tags:
      - orders
  /orders/all_orders:
    get:
      description: Get a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.OrderDoc'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all orders
      tags:
      - orders
  /orders/assign:
    post:
      description: Assign nearest available drivers to pending orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Run auto-assignment for pending orders
      tags:
      - orders
  /orders/by-customer/{customer_id}:
    get:
      description: Fetch order(s) using Customer ID
      parameters:
      - description: Customer ID
        in: path
        name: customer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/order.OrderDoc'
            type: array
        "400":
          description: Invalid Customer ID
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get order by Customer ID
      tags:
      - orders
  /orders/by-id/{id}:
    get:
      description: Fetch a single order using UUID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/order.OrderDoc'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
  /orders/create:
    post:
      consumes:
      - application/json
      description: Creates an order and returns the new object
      parameters:
      - description: Order input
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.CreateOrderRequestDoc'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/order.OrderDoc'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /payments/{id}:
    get:
      description: Fetch a single payment using payment ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.Payment'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get payment by ID
      tags:
      - payments
  /payments/{order_id}:
    get:
      description: Fetch payment(s) using Order ID
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payment.Payment'
            type: array
        "400":
          description: Invalid Order ID
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get payment by Order ID
      tags:
      - payments
  /payments/all_payments:
    get:
      description: Get a list of all payments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payment.Payment'
            type: array
      security:
      - BearerAuth: []
      summary: List all payments
      tags:
      - payments
  /payments/create:
    post:
      consumes:
      - application/json
      description: Register new payment with order_id, amount, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/payment.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payment.Payment'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to create payment
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create new payment
      tags:
      - payments
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"